SmartCropAdvisory/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0
psycopg2-binary==2.9.7
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
Pillow==10.0.0
